pyproject.toml
[tool.poetry]
name = "openai-playground"
version = "0.1.0"
description = ""
authors = ["Weiler, Ryan <ryanweiler92@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.57.2"
pydantic = "^2.10.3"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
huggingface-hub = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
